+++ proxy/testing/test_contract_reverting.py
@@ -1,25 +1,16 @@
 import unittest
 import os
-import requests
 
 from web3 import exceptions as web3_exceptions
 from solana.rpc.api import Client as SolanaClient
 from eth_account.account import LocalAccount
 
-from .testing_helpers import SolidityContractDeployer
+from .testing_helpers import SolidityContractDeployer, request_airdrop
 from ..common_neon.emulator_interactor import decode_revert_message
 
 from solcx import install_solc
 install_solc(version='0.7.6')
 
-def request_airdrop(address):
-    url = 'http://faucet:3333/request_neon'
-    data = '{"wallet": "' + address + '", "amount": 5}'
-    r = requests.post(url, data=data)
-    if not r.ok:
-        print()
-        print('Bad response:', r)
-    assert(r.ok)
 
 class TestContractReverting(unittest.TestCase):
