+++ proxy/testing/testing_helpers.py
@@ -1,5 +1,6 @@
 import dataclasses
 import os
+import requests
 import solcx
 from eth_account.account import LocalAccount
 from web3 import Web3, eth as web3_eth
@@ -45,3 +46,15 @@ class SolidityContractDeployer:
     @property
     def web3(self) -> Web3:
         return self._web3
+
+
+def request_airdrop(address):
+    FAUCET_URL = os.environ.get('FAUCET_URL', 'http://faucet')
+    FAUCET_PORT = os.environ.get('FAUCET_PORT', '3333')
+    url = FAUCET_URL + ':' + FAUCET_PORT + '/request_neon'
+    data = '{"wallet": "' + address + '", "amount": 5}'
+    r = requests.post(url, data=data)
+    if not r.ok:
+        print()
+        print('Bad response:', r)
+    assert(r.ok)
